requirements.txt
altgraph==0.17
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.0.1
ffmpeg==1.4
future==0.18.2
idna==2.10
importlib-metadata==3.7.3
openpyxl==3.0.7
pefile==2019.4.18
psycopg2==2.8.6
Pybooru==4.2.2
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
python-telegram-bot==13.3
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
xlrd==2.0.1
zipp==3.4.1
